Better WHATWG BIG5 round-tripping.
WHATWG specifies that certain BIG5 codes be decoded to two-codepoint
graphemes, but it doesn't specify converting them back when encoding.
This makes it possible to successfully decode text that can then
not be properly re-encoded (even lossily).

This commit adds support to handle those graphemes specially in the
encoder as well, allowing for round-tripping of those BIG5 codes.